Various bugfix
Former-commit-id: 7f95263041d1859eaac30ce8745fabf1b39ab467 [formerly 90d60c2fa9490fbfe850a0ce967c16e58dafe136] [formerly 49455b379a458cb773bbfe381dbb8bdff4ccd25c [formerly d6bca6e3a75fe42ec56a27a3ef6d574f42f536e3] [formerly 7c32cf5ae49938f521643c88be48be6c692c99a9] [formerly 5233293b8d8c6627979c57778fa91ddc7a0f36a7 [formerly d4bbb1960666b9e99bc4325e3116ac2f92e35df8]] [formerly 49455b379a458cb773bbfe381dbb8bdff4ccd25c [formerly d6bca6e3a75fe42ec56a27a3ef6d574f42f536e3] [formerly 7c32cf5ae49938f521643c88be48be6c692c99a9] [formerly 5233293b8d8c6627979c57778fa91ddc7a0f36a7 [formerly d4bbb1960666b9e99bc4325e3116ac2f92e35df8]] [formerly c99fa5fb5c3d8841a43094cd058f70d0a2090c29 [formerly 83d4410a6ea863de80929830e3ee519734d730c2]]]]
Former-commit-id: 1c297c6f646ffa4b4bf52e00d48eb363dbdef124
Former-commit-id: cb5c5f8837b864c1bfcb3ef5f6f04714e910bf38
Former-commit-id: ae93c24bfe20e7acf427dc84cfed896ecbc56c3e
Former-commit-id: 2e4f11766453c2ad9fd3b2d3482f524d2aff4a86
Former-commit-id: 831a1e2fb11decad0279b1f83bf6e8fa6a33a9ae [formerly 44dda0d80bc12070f2b1ac84e45ba085ed69f0ba]
Former-commit-id: c6dbd8a914bc420bd8ad583ab31410f82d7735d7
Former-commit-id: 30fa73dddb82d2229b7d6e06737b3d8c4c894376

+++ Source/Raw.Net/Source/Decoder/OrfDecoder.cs
@@ -163,7 +163,7 @@ namespace RawNet
                             bits.SkipBitsNoFill((uint)high + 1 + 3);
                         }
 
-                        acarry0[0] = (high << nbits) | bits.GetBits((uint)nbits);
+                        acarry0[0] = (uint)(high << nbits) | bits.GetBits((uint)nbits);
                         diff = (int)((acarry0[0] ^ sign) + acarry0[1]);
                         acarry0[1] = (diff * 3 + acarry0[1]) >> 5;
                         acarry0[2] = acarry0[0] > 16 ? 0 : acarry0[2] + 1;
@@ -230,7 +230,7 @@ namespace RawNet
                             bits.SkipBitsNoFill((uint)high + 1 + 3);
                         }
 
-                        acarry1[0] = (high << nbits) | bits.GetBits((uint)nbits);
+                        acarry1[0] = (uint)(high << nbits) | bits.GetBits((uint)nbits);
                         diff = (int)((acarry1[0] ^ sign) + acarry1[1]);
                         acarry1[1] = (diff * 3 + acarry1[1]) >> 5;
                         acarry1[2] = acarry1[0] > 16 ? 0 : acarry1[2] + 1;
